Draggable Weapon Card System
User Story:
As a player, I want to have Weapons as UI cards so that I know what weapons I have in reserve in case I lose Segments.

AND

As a player, I want to drag Weapon cards onto empty Segments so that I can attach weapons to my Centipede to fight enemies.

Description:
Previously, Weapons were pickups that were laid across the ground and be immediately attached to the Centipede upon collision. However, a core aspect of the game was to have draggable cards. With these User Stories, and this commit, Players are able to drag and attach Weapons onto empty Segments.

Michael Wu