Ant Orientations and Models
User Stories:

As a hunter ant I want to have a model similar to farmers, but larger and a special location for weapons, so that I can be differentiated from the other types of ants.
AND
As a guard ant I want to have a model slightly larger than worker ants and have a significantly larger head, so that I can be differentiated from the other types of ants.

Description:
Implemented the models for the hunter and guard ants