First Centipede Support
User Story:
As a player, I want to have Weapons as UI cards so that I know what weapons I have in reserve in case I lose Segments.

AND

As a player, I want to drag Weapon cards onto empty Segments so that I can attach weapons to my Centipede to fight enemies.

AND

As a player, I want my view of the Centipede to be unobstructed so that I can always see my Centipede and where it is heading.

Description:
The Centipede used for the tech demo was incompatible with the Spring Arm and the Draggable Weapon Cards.
This has been fixed.

Michael Wu