Spring Arm Component Rotation Bug Fix
User Story:
As a player, I want my view of the Centipede to be unobstructed so that I can always see my Centipede and where it is heading.

Description:
When switching between Inherit Rotation settings, the component would break if it was Using Lag. This has been fixed.

Michael Wu