Segment Alignment
User Story:
As a player, I want the centipede's feet to crawl like a real centipede so that the centipede has a movement animation.
AND
As a player, I want the centipede's feet to adapt to the environment so that the legs are positioned correctly and appears more realistic.

Description:
Fixed a long-standing issue where Segments would not realign themselves with their ForwardNeighbour, often appearing upside-down.

Michael Wu