FABRIK Optimisations
User Story:
As a player, I want the centipede's feet to crawl like a real centipede so that the centipede has a movement animation.
AND
As a player, I want the centipede's feet to adapt to the environment so that the legs are positioned correctly and appears more realistic.

Description:
Optimisations on FABRIK include ignoring execution if:
The FABRIK GameObject is too far from the Camera.
OR
The Spring Arm Component has a large Distance set, such that execution is unnecessary.

Calculating the Distance from the Camera every frame may be too expensive as well - needing to Deproject a world point to a screen point. This is not implemented.

Michael Wu