FABRIK Centipede in Environment Test Scene
User Story:
As a player, I want the centipede's feet to crawl like a real centipede so that the centipede has a movement animation.
AND
As a player, I want the centipede's feet to adapt to the environment so that the legs are positioned correctly and appears more realistic.

Description:
A FABRIK enabled Centipede is now in the Environment Test Scene.

Michael Wu