Walking on Walls > 40 Degrees
User Story:
As a player, I want the centipede to be able to walk on walls and the ceiling, so that I have more freedom when moving.

Description:
Previously this was tested and only worked on walls angled < 40 degrees. This commit should **hopefully** allow the Centipede to walk on any wall, provided the delta degrees from one wall to the next is not too steep.

Michael Wu.