Improved Segment Alignment
User Story:
As a player, I want the centipede's feet to crawl like a real centipede so that the centipede has a movement animation.

AND

As a player, I want the centipede's feet to adapt to the environment so that the legs are positioned correctly and appears more realistic.

Description:
Segments should now at least try to align themselves with the terrain underneath them. Otherwise, they will align themselves with their ForwardNeighbour.

Michael Wu