Pause Menu & Inherit Rotation
User Stories:
As a player, I want to see and control my Centipede in a third-person perspective so that I can see what is ahead of me.

AND

As a player, I want to be able to change my Camera follow settings so that the game is better fit to my playstyle.

Description:
The Player is now able to change the Camera settings into either 3rd person, or Default Global.

Also, the ability to toggle Lag has been removed because it breaks when Inherit Rotation = true.

Michael Wu