Right-Mouse Camera Drag
User Story:
As a player I want the camera to be adjustable with the right mouse so that I have a better view while moving.

Description:
Dragging the mouse while right clicking will cause the game Camera to Orbit the Centipede accordingly. Only works when the Spring Arm is NOT Inheriting the Target's Rotation.

Michael Wu.