AI Balancing
User Stories:
As any ant I want my starting health to be increased, while still being based on the same ratios, so that I can appear more challenging to players by being harder to kill
AND
As any ant, I want to, when taking damage, only sometimes go into the damage state, so that players cannot abuse this temporarily stopping of my attack.
AND
As a guard ant, I do not want to be able to call for backup, so that I can be easier to defeat due to my overall massive toughness.

Description
Did some overall refinement and balancing to the various stats of the AI as shown through the above user stories