Dynamic Centipede-Segment-Terrain Alignment
An issue exists where the Centipede's Head and the Segments will remain unnaturally rotated after turning on banked Terrain and then going onto a relatively 'flat' Terrain, without stopping to let the Head and Segments to realign themselves.

Previously, the Head and Segments would only align themselves when the main Centipede Body is stopped. This was for smoothing the realignment of the Head and Segments.

Now, when any individual Head or Segment is rotated at an angle > kErrorAngle, relative to the below Normal, a realignment will be enforced.