Added Collectible Parent Parts
User Stories:
As a player I want all of the parts to be wrapped in cobwebs, so that I can easily distinguish them from the rest of the map

AND

As a player, I want to have an option to collect all my parents' parts, so that I can have an optional secondary objective so that exploration is more worthwhile

Description:
Currently, move to the parts location and break the cobweb to release the part and auto collect it. 14 different ones will exist around the map.