Weapon Fire Rate Bug Fixes
The Fire Rate of a Weapon would be continuously updated, even without a shot fired. This meant that Weapons with a low Fire Rate, e.g., Launcher - 1, can only Fire at 1-second *absolute* intervals.

I.e., If a Launcher was attached 2.75 seconds into the game and never fired, the Launcher would not be able to Fire for another .25 seconds, when it should.